requirements.txt
absl-py==2.1.0
APScheduler==3.10.4
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.55.6
idna==3.10
itsdangerous==2.2.0
jax==0.4.38
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml-dtypes==0.5.2
numpy==1.26.4
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt-einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.8
psycopg2-binary==2.9.10
pycparser==2.22
pygame==2.6.1
pyparsing==3.2.1
pyTelegramBotAPI==4.24.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyvirtualcam==0.12.0
requests==2.32.3
scipy==1.14.1
simplejpeg==1.7.6
six==1.17.0
sounddevice==0.5.1
transliterate==1.10.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
dlib==19.24.6
cmake==3.31.4